pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "Agent Bazaar"
authors = ["Kenan Agyel <kenanagyel@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
pydantic = { extras = ["email"], version = "^2.6.1" }
fastapi = "^0.109.1"
uvicorn = "^0.27.0"
uvloop = "^0.19.0"
httptools = "^0.6.1"
uuid = "^1.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
SQLAlchemy-Utils = "^0.41.1"
python-jose = "^3.3.0"
SQLAlchemy = "^2.0.25"
pytest = "^7.4.2"
python-multipart = "^0.0.9"
greenlet = "^2.0.2"
httpx = "^0.26.0"
pydantic-settings = "^2.0.3"
gunicorn = "^22.0.0"
bcrypt = "^4.1.1"
fastcrud = "^0.12.0"
faker = "^26.0.0"
psycopg2-binary = "^2.9.9"
pytest-mock = "^3.14.0"
streamlit = "^1.39.0"
requests = "^2.32.3"
pyyaml = "^6.0.2"
isort = "^5.13.2"
black = "^24.10.0"
docker = "^7.1.0"
pytest-asyncio = "^0.23.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"